api/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
etherscan-python==2.1.0
Flask==2.0.0
Flask-Cors==3.0.10
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
numpy==1.20.3
pandas==1.2.4
pycoingecko==2.0.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-cache==0.6.3
six==1.16.0
url-normalize==1.4.3
urllib3==1.26.4
Werkzeug==2.0.0
